What I made for FileIO is good enough for now.
To prevent that the file will be overwritten you should change the offSet with File.seekTo(). You can also add a newline to the string so that it will be placed on the next line. I haven't implemented this yet, because I don't know if I'll need it.

If you create a DictionaryFile struct and declare it as undefined, the initialized bool is also undefined. This means that the DictionaryFileError is not returned.