File io (#1)
* Initial setup FileIO

Might consider to create a struct that saves directory and file name.

* DictionaryFile struct

DictionaryFile.write_file does not create a file when it doesn't exist.

There is also no new directory created if the given directory name doesn't exist.

* What I made for FileIO is good enough for now.

To prevent that the file will be overwritten you should change the offSet with File.seekTo(). You can also add a newline to the string so that it will be placed on the next line. I haven't implemented this yet, because I don't know if I'll need it.

If you create a DictionaryFile struct and declare it as undefined, the initialized bool is also undefined. This means that the DictionaryFileError is not returned.